refactor(backup): move getBackupTimestamp utility to a shared location for consistency
